moving to int64 fib

+++ source/tests/fiblib-test/fibonacci_test.cpp
@@ -1,4 +1,5 @@
 
+#include <stdint.h>
 #include <gmock/gmock.h>
 
 
@@ -13,9 +14,9 @@ TEST_F(fibonacci_test, CheckSomeResults)
 {
     fiblib::Fibonacci fib;
 
-    EXPECT_EQ((unsigned int) 0, fib(0));
-    EXPECT_EQ((unsigned int) 1, fib(1));
-    EXPECT_EQ((unsigned int) 1, fib(2));
-    EXPECT_EQ((unsigned int)21, fib(8));
+    EXPECT_EQ((uint64_t) 0, fib(0));
+    EXPECT_EQ((uint64_t) 1, fib(1));
+    EXPECT_EQ((uint64_t) 1, fib(2));
+    EXPECT_EQ((uint64_t)21, fib(8));
     // ...
 }
